- remove dagger2
remove dependencey injection from this branch
