+++ .idea/modules.xml
@@ -3,7 +3,10 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/NewsFeedapp.iml" filepath="$PROJECT_DIR$/NewsFeedapp.iml" />
+      <module fileurl="file://$PROJECT_DIR$/NewsFeedapp-main.iml" filepath="$PROJECT_DIR$/NewsFeedapp-main.iml" />
       <module fileurl="file://$PROJECT_DIR$/aboutme/aboutme.iml" filepath="$PROJECT_DIR$/aboutme/aboutme.iml" />
+      <module fileurl="file://$PROJECT_DIR$/aboutme/aboutme.iml" filepath="$PROJECT_DIR$/aboutme/aboutme.iml" />
+      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
