+++ app/build.gradle
@@ -82,10 +82,6 @@ dependencies {
     implementation 'io.reactivex.rxjava2:rxandroid:2.0.1'
     implementation 'com.jakewharton.retrofit:retrofit2-rxjava2-adapter:1.0.0'
 
-    def dagger_version = "2.17"
-    implementation "com.google.dagger:dagger:$dagger_version"
-    annotationProcessor "com.google.dagger:dagger-compiler:$dagger_version"
-
     implementation 'android.arch.persistence.room:runtime:1.1.1'
     annotationProcessor "android.arch.persistence.room:compiler:1.1.1"
     implementation 'android.arch.lifecycle:extensions:1.1.1'
@@ -94,13 +90,6 @@ dependencies {
     implementation 'com.github.drawers:SpinnerDatePicker:1.0.6'
 
     implementation 'com.squareup.picasso:picasso:2.71828'
-
-    implementation 'com.google.code.findbugs:jsr305:3.0.2' // to find dagger2 bugs
-
-    implementation 'com.google.auto.factory:auto-factory:1.0-beta3'
-    annotationProcessor 'com.google.auto.factory:auto-factory:1.0-beta3'
-
-    implementation 'javax.annotation:jsr250-api:1.0' // to fix cannot find symbol class Generated
 
     //adding paging
     def paging_version = "1.0.1"
