--- app/src/main/java/com/gmail/tarekmabdallah91/news/di/interfaces/ApplicationContext.java
@@ -1,29 +1,0 @@
-/*
- *
- * Copyright 2019 tarekmabdallah91@gmail.com
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-package com.gmail.tarekmabdallah91.news.di.interfaces;
-
-import javax.inject.Qualifier;
-
-/**
- * Created by Hari on 17/12/17.
- */
-@Qualifier
-public @interface ApplicationContext {
-
-}
