--- app/src/main/java/com/gmail/tarekmabdallah91/news/di/modules/ActivityModule.java
@@ -1,51 +1,0 @@
-/*
- *
- * Copyright 2019 tarekmabdallah91@gmail.com
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-package com.gmail.tarekmabdallah91.news.di.modules;
-
-import android.app.Activity;
-import android.content.Context;
-
-import com.gmail.tarekmabdallah91.news.di.interfaces.ApplicationScope;
-
-import javax.inject.Named;
-
-import dagger.Module;
-import dagger.Provides;
-
-/**
- * Created by Hari on 23/11/17.
- */
-@Module
-public class ActivityModule {
-
-    private final Context context;
-
-    ActivityModule(Activity context){
-        this.context = context;
-    }
-
-    /*
-     * @Named — which is used to differentiate dependency providers.
-     * @Qualifier annotation, an alternate to @Named annotation.
-     */
-    @Named("activity_context")
-    @ApplicationScope
-    @Provides
-    public Context context(){ return context; }
-}
