--- app/src/main/java/com/gmail/tarekmabdallah91/news/di/modules/ApplicationModule.java
@@ -1,56 +1,0 @@
-/*
- *
- * Copyright 2019 tarekmabdallah91@gmail.com
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-package com.gmail.tarekmabdallah91.news.di.modules;
-
-import android.app.Application;
-import android.content.Context;
-import android.content.SharedPreferences;
-
-import javax.inject.Singleton;
-
-import dagger.Module;
-import dagger.Provides;
-
-import static com.gmail.tarekmabdallah91.news.utils.Constants.SHARED_PREF;
-
-@Module
-public class ApplicationModule {
-
-    private Application application;
-
-    public ApplicationModule(Application application) {
-        this.application = application;
-    }
-
-    @Provides
-    @Singleton
-    Application provideApplication() {
-        return application;
-    }
-
-    @Provides
-    Context provideContext() {
-        return application;
-    }
-
-    @Provides
-    SharedPreferences provideSharedPrefs() {
-        return application.getSharedPreferences(SHARED_PREF, Context.MODE_PRIVATE);
-    }
-}
