--- app/src/main/java/com/gmail/tarekmabdallah91/news/di/modules/ContextModule.java
@@ -1,49 +1,0 @@
-/*
- *
- * Copyright 2019 tarekmabdallah91@gmail.com
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-package com.gmail.tarekmabdallah91.news.di.modules;
-
-import android.content.Context;
-
-import com.gmail.tarekmabdallah91.news.di.interfaces.ApplicationContext;
-import com.gmail.tarekmabdallah91.news.di.interfaces.ApplicationScope;
-
-import dagger.Module;
-import dagger.Provides;
-
-/**
- * Created by Hari on 23/11/17.
- */
-@Module
-public class ContextModule {
-
-    private Context context;
-
-    public ContextModule(Context context){
-        this.context = context;
-    }
-
-    /*
-     * @Named — which is used to differentiate dependency providers.
-     * @Qualifier annotation, an alternate to @Named annotation.
-     */
-    @ApplicationContext
-    @ApplicationScope
-    @Provides
-    public Context context(){ return context.getApplicationContext(); }
-}
