--- app/src/main/java/com/gmail/tarekmabdallah91/news/utils/GsonObject.java
@@ -1,32 +1,0 @@
-/*
- *
- * Copyright 2019 tarekmabdallah91@gmail.com
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-package com.gmail.tarekmabdallah91.news.utils;
-
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-
-public class GsonObject {
-
-    private static Gson gson; //http://www.javacreed.com/simple-gson-example/
-
-    public static Gson getGsonInstance() {
-        if (null == gson) gson = new GsonBuilder().create();
-        return gson;
-    }
-}
