--- app/src/main/java/com/gmail/tarekmabdallah91/news/views/section/di/SectionActivityComponent.java
@@ -1,31 +1,0 @@
-/*
- *
- * Copyright 2019 tarekmabdallah91@gmail.com
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-package com.gmail.tarekmabdallah91.news.views.section.di;
-
-import com.gmail.tarekmabdallah91.news.views.articlesFragment.ArticlesFragment;
-
-import javax.inject.Singleton;
-
-import dagger.Component;
-
-@Singleton
-@Component(modules = {SectionActivityModule.class})
-public interface SectionActivityComponent {
-    ArticlesFragment getArticlesFragment();
-}
